Mark up blog posts with h-entry/h-feed.
Specifically this marks up entries with the following fields:

 - h-feed (on index page):
   - nested h-entry objects
 - h-entry (in index h-feed and on post pages):
   - p-name
   - p-summary
   - e-content (on post pages)
   - dt-published
   - p-author (if author field present, see below)
     - h-card
   - u-url (in index h-feed)

p-author
--------

Allows for authors other than myself with two new optional post
metadata fields:

 - author
 - authorurl

If an author is specified, an authorurl must also be. If an author
is specified, the h-entrys regarding this post contain a p-author
field (with a h-card and rel=author link).